requirements.txt
torchtune>=0.4.0
torchao>=0.9.0
soundfile
vector_quantize_pytorch
transformers>=4.45.0
einops>=0.8.0
safetensors>=0.4.0
sentencepiece>=0.2.0
accelerate>=1.0.0
numpy
protobuf
omegaconf
